Add row by row CSV reader and fixes
Resolves the requested changes from #30:

- Split the logic of the common parts of the CSV parser to two
  separated packages: csvparse and awsparser.
  The main aim of the csvparse package is to provide usefull generic
  API for reading the CSV files. awsparser is more specific package
  to parse the AWS Cost And Usage Reports files.

- Add CSV row-by-row reader

- Add AWS specific row-by-row reader

See pull request #30
